LPS-77456 Move WikiImporterTracker to wiki-web.
