+++ modules/apps/collaboration/wiki/wiki-web/src/main/resources/META-INF/resources/wiki/init.jsp
@@ -41,7 +41,6 @@ page import="com.liferay.wiki.exception.PageTitleException" %><%@
 page import="com.liferay.wiki.exception.PageVersionException" %><%@
 page import="com.liferay.wiki.exception.RequiredNodeException" %><%@
 page import="com.liferay.wiki.exception.WikiFormatException" %><%@
-page import="com.liferay.wiki.importer.impl.WikiImporterTracker" %><%@
 page import="com.liferay.wiki.social.WikiActivityKeys" %><%@
 page import="com.liferay.wiki.util.WikiPageAttachmentsUtil" %><%@
 page import="com.liferay.wiki.util.WikiUtil" %><%@
@@ -55,6 +54,7 @@ page import="com.liferay.wiki.web.internal.display.context.logic.WikiVisualizati
 page import="com.liferay.wiki.web.internal.display.context.util.WikiRequestHelper" %><%@
 page import="com.liferay.wiki.web.internal.display.context.util.WikiSocialActivityHelper" %><%@
 page import="com.liferay.wiki.web.internal.display.context.util.WikiURLHelper" %><%@
+page import="com.liferay.wiki.web.internal.importer.WikiImporterTracker" %><%@
 page import="com.liferay.wiki.web.internal.portlet.toolbar.item.WikiPortletToolbarContributor" %><%@
 page import="com.liferay.wiki.web.internal.security.permission.resource.WikiNodePermission" %><%@
 page import="com.liferay.wiki.web.internal.security.permission.resource.WikiPagePermission" %><%@
